requirements.txt
requests
python-dotenv
aiohttp
smolagents

